backend: added user account support and authentication and authorization on requests

+++ backend/src/main/java/gang/gutscheingang/controllers/UserController.java
@@ -11,57 +11,59 @@ import gang.gutscheingang.repositories.VoucherRepository;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
+import org.springframework.security.core.Authentication;
 
 import javax.validation.Valid;
+import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
 
 @RestController
 @RequestMapping("user")
 @CrossOrigin(origins = "*", allowedHeaders = "*")
-@Tag(name="Users")
+@Tag(name = "Users")
 public class UserController extends GenericController {
 
     private UserRepository userRepository;
     private CompanyRepository companyRepository;
     private VoucherRepository voucherRepository;
+    private BCryptPasswordEncoder bCryptPasswordEncoder;
 
     @Autowired
-    public UserController(UserRepository userRepository, CompanyRepository companyRepository, VoucherRepository voucherRepository) {
+    public UserController(UserRepository userRepository, CompanyRepository companyRepository, VoucherRepository voucherRepository, BCryptPasswordEncoder bCryptPasswordEncoder) {
         this.userRepository = userRepository;
         this.companyRepository = companyRepository;
         this.voucherRepository = voucherRepository;
+        this.bCryptPasswordEncoder = bCryptPasswordEncoder;
     }
 
     @PostMapping(produces = "application/json")
     public SystemUser createUser(@RequestBody @Valid SystemUser systemUser) {
-
+        systemUser.setPassword(bCryptPasswordEncoder.encode(systemUser.getPassword()));
+        systemUser.setUsername(systemUser.getUsername().toLowerCase());
         return userRepository.save(systemUser);
     }
 
-    @PostMapping(value = "/{uuid}/voucher", produces = "application/json")
-    public Voucher buyVoucher(@PathVariable UUID uuid, @RequestBody @Valid VoucherBuyTransaction transaction) {
-        SystemUser systemUser = userRepository.findById(uuid).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
-        Company company = companyRepository.findById(transaction.getCompanyUuid()).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));;
-        Voucher voucher = systemUser.buyVoucher(company, transaction.getValueInEurCt());
-        return voucherRepository.save(voucher);
-    }
-
-
-    @GetMapping(value = "/{uuid}/companies", produces = "application/json")
-    public List<Company> getCompaniesOfUser(@PathVariable UUID uuid) {
-        return userRepository.findById(uuid).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND)).getCompanyList();
-    }
-
     @GetMapping(value = "/{uuid}", produces = "application/json")
-    public SystemUser getUserById(@PathVariable UUID uuid) {
+    public SystemUser getUserById(@PathVariable UUID uuid, Principal principal) {
+        SystemUser currentUser = userRepository.findByEmailIgnoreCase(principal.getName());
+        if (!currentUser.getUuid().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         return userRepository.findByUuid(uuid);
     }
 
     @PutMapping(value = "/{uuid}", produces = "application/json")
-    public SystemUser updateUser(@PathVariable UUID uuid, @RequestBody @Valid SystemUser systemUser) {
+    public SystemUser updateUser(@PathVariable UUID uuid, @RequestBody @Valid SystemUser systemUser, Principal principal) {
+        SystemUser currentUser = userRepository.findByEmailIgnoreCase(principal.getName());
+        if (!currentUser.getUuid().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         return userRepository.findById(uuid)
                 .map(
                         foundSystemUser -> {
@@ -71,12 +73,56 @@ public class UserController extends GenericController {
     }
 
     @DeleteMapping(value = "/{uuid}", produces = "application/json")
-    public void deleteUser(@PathVariable UUID uuid) {
+    public void deleteUser(@PathVariable UUID uuid, Principal principal) {
+        SystemUser currentUser = userRepository.findByEmailIgnoreCase(principal.getName());
+        if (!currentUser.getUuid().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         try {
             userRepository.deleteById(uuid);
 
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         }
+    }
+
+    @PostMapping(value = "/{uuid}/voucher", produces = "application/json")
+    public Voucher buyVoucher(@PathVariable UUID uuid, @RequestBody @Valid VoucherBuyTransaction transaction, Principal principal) {
+        SystemUser currentUser = userRepository.findByEmailIgnoreCase(principal.getName());
+        if (!currentUser.getUuid().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+        SystemUser systemUser = userRepository.findById(uuid).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Company company = companyRepository.findById(transaction.getCompanyUuid()).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        Voucher voucher = systemUser.buyVoucher(company, transaction.getValueInEurCt());
+        voucher = voucherRepository.save(voucher);
+        userRepository.save(systemUser);
+
+        company.addVoucher(voucher);
+        companyRepository.save(company);
+        return voucher;
+    }
+
+    @GetMapping(value = "/{uuid}/companies", produces = "application/json")
+    public List<Company> getCompaniesOfUser(@PathVariable UUID uuid, Principal principal) {
+        SystemUser currentUser = userRepository.findByEmailIgnoreCase(principal.getName());
+        if (!currentUser.getUuid().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+        return userRepository.findById(uuid).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND)).getCompanyList();
+    }
+
+    @GetMapping(value = "/{uuid}/vouchers", produces = "application/json")
+    public List<Voucher> getVouchersOfUser(@PathVariable UUID uuid, Principal principal) {
+        SystemUser currentUser = userRepository.findByEmailIgnoreCase(principal.getName());
+        if (!currentUser.getUuid().equals(uuid)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
+        return userRepository.findById(uuid).orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND)).getVoucherList();
     }
 }
